deduplicate read_proof_check
