+++ pallet-ismp/primitives/state-machine/Cargo.toml
@@ -21,7 +21,7 @@ scale-info = { version = "2.1.1", default-features = false, features = ["derive"
 frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
 sp-core = { default-features = false, git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
 sp-trie = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-
+hash-db = { version = "0.16.0", default-features = false }
 
 ismp-primitives = { path = "..", default-features = false }
 pallet-ismp = { path = "../..", default-features = false }
@@ -41,5 +41,6 @@ std = [
     "sp-core/std",
     "pallet-ismp/std",
     "ismp-primitives/std",
-    "sp-trie/std"
+    "sp-trie/std",
+    "hash-db/std"
 ]
