+++ parachain/src/consensus.rs
@@ -39,8 +39,8 @@ use sp_runtime::{
     traits::{BlakeTwo256, Header as _},
     DigestItem,
 };
-use sp_trie::{HashDBT, LayoutV0, StorageProof, Trie, TrieDBBuilder, EMPTY_PREFIX};
-use substrate_state_machine::SubstrateStateMachine;
+use sp_trie::StorageProof;
+use substrate_state_machine::{read_proof_check, SubstrateStateMachine};
 
 use crate::RelayChainOracle;
 
@@ -216,36 +216,4 @@ pub fn parachain_header_storage_key(para_id: u32) -> StorageKey {
     storage_key.extend_from_slice(sp_io::hashing::twox_64(&encoded_para_id).as_slice());
     storage_key.extend_from_slice(&encoded_para_id);
     StorageKey(storage_key)
-}
-
-/// Lifted directly from [`sp_state_machine::read_proof_check`](https://github.com/paritytech/substrate/blob/b27c470eaff379f512d1dec052aff5d551ed3b03/primitives/state-machine/src/lib.rs#L1075-L1094)
-pub fn read_proof_check<H, I>(
-    root: &H::Out,
-    proof: StorageProof,
-    keys: I,
-) -> Result<BTreeMap<Vec<u8>, Option<Vec<u8>>>, Error>
-where
-    H: hash_db::Hasher,
-    H::Out: Debug,
-    I: IntoIterator,
-    I::Item: AsRef<[u8]>,
-{
-    let db = proof.into_memory_db();
-
-    if !db.contains(root, EMPTY_PREFIX) {
-        Err(Error::ImplementationSpecific("Invalid Proof".into()))?
-    }
-
-    let trie = TrieDBBuilder::<LayoutV0<H>>::new(&db, root).build();
-    let mut result = BTreeMap::new();
-
-    for key in keys.into_iter() {
-        let value = trie
-            .get(key.as_ref())
-            .map_err(|e| Error::ImplementationSpecific(format!("Error reading from trie: {e:?}")))?
-            .and_then(|val| Decode::decode(&mut &val[..]).ok());
-        result.insert(key.as_ref().to_vec(), value);
-    }
-
-    Ok(result)
 }
